+++ naga/src/back/spv/mod.rs
@@ -1191,4 +1191,6 @@ pub fn supported_capabilities() -> crate::valid::Capabilities {
         | Caps::PER_VERTEX
         // No RAY_TRACING_PIPELINE
         | Caps::DRAW_INDEX
+        | Caps::MEMORY_DECORATION_COHERENT
+        | Caps::MEMORY_DECORATION_VOLATILE
 }