+++ wgpu-hal/src/dx12/adapter.rs
@@ -490,7 +490,8 @@ impl super::Adapter {
             | wgt::Features::FLOAT32_FILTERABLE
             | wgt::Features::TEXTURE_ATOMIC
             | wgt::Features::PASSTHROUGH_SHADERS
-            | wgt::Features::EXTERNAL_TEXTURE;
+            | wgt::Features::EXTERNAL_TEXTURE
+            | wgt::Features::MEMORY_DECORATION_COHERENT;
 
         //TODO: in order to expose this, we need to run a compute shader
         // that extract the necessary statistics out of the D3D12 result.